Working with CSV
A CSV file stores tabular data (numbers and text) in plain text form. A CSV file consists of any number of records, separated by line breaks of some kind. Each record consists of fields, separated by a literal comma. In some regions, the separator might be a semi-colon.

There are several ways to load a CSV file in Python. We are going to explore all the methods through which we can load CSV files, and we will also cover the parameters of CSV files.

Additionally, please note that all the files used in the notebook are included in this folder.